Solucionado hpta

+++ src/main/java/com/usta/proyectofinal/models/dao/RolDao.java
@@ -3,5 +3,5 @@ package com.usta.proyectofinal.models.dao;
 import com.usta.proyectofinal.entities.RolEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RolDao  extends CrudRepository<RolEntity,Integer> {
+public interface RolDao  extends CrudRepository<RolEntity,Long> {
 }
